backend/requirements.txt
Django>=3.2
psycopg2-binary>=2.8
django-cors-headers
djangorestframework>=3.14.0